requirements.txt
mediapipe>=0.10.13
opencv-python>=4.8.1
pyautogui>=0.9.54
pyyaml>=6.0
bleak>=0.21.1
numpy>=1.24.3
pytest>=7.4.0
